// src/app/core/data.enum.ts
export enum DATA {

    TOKEN = "token",
    USERNAME = "username",
    USERID = "userId",
    ROLE = "role",
    LAST_ACTION = "lastAction",
    FIRST = "first",
    TIME_TO_EXPIRE = "timeToExpire",
    ACCOUNT_EXPIRED = "accountExpired",
    PASSWORD_ABOUT_TO_EXPIRE = "passwordAboutToExpire",
    APP_SUBJECT = "App-Subject",
    AUTHORIZATION = "Authorization",
    BEARER = "Bearer ",
    CONTENT_TYPE = "Content-Type",
    IS_ACCOUNT_EXPIRED = "isAccountExpired",
    CACHE_CONTROL = "cache-control"

}